refactor: move Button to src/components/atom
