+++ src/components/options/OptionsTemplate.tsx
@@ -4,7 +4,7 @@ import { Container, Columns } from "react-bulma-components"
 import Header from "../header"
 import ItemList, { Props as ItemListProps } from "../itemList"
 import ItemEditor, { Props as ItemEditorProps } from "../itemEditor"
-import { Button } from "../Button"
+import { Button } from "../atom/Button"
 import { CustomBlock } from "../../domain/customBlock/block"
 import { BlockEditorProps, BlockEditor } from "../editor/BlockEditor"
 import { Item } from "../../domain/item/item"
